# api.yml
# This is the OpenAPI specification for your service. It is used to generate the client code.
# To generate the server code, use the Gradle task `openApiGenerate`. The 
# generated code will be placed in `build/generate-resources/main`.
#
# We will evaluate you based on how well you design your API.
# 1. How well it follows REST principles.
# 2. How easy it is to understand and use.
#
# References
# 1. https://swagger.io/specification/
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  license:
    name: MIT
servers:
  - url: http://localhost
paths:
  /v1/user/login:
    post:
      operationId: UserLogin
      summary: Existing user login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '400':
          description: Bad request - Invalid input
        '500':
          description: Internal server error
  /v1/user:
    get:
      operationId: GetUser
      summary: Get an existing new user
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserResponse'
        '403':
          description: Forbidden
        '500':
          description: Internal server error
    post:
      operationId: RegisterUser
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterUserResponse'
        '400':
          description: Bad request - Invalid input
        '409':
          description: Conflict
        '500':
          description: Internal server error
    put:
      operationId: UpdateUser
      summary: Update an existing user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateUserResponse'
        '403':
          description: Forbidden
        '409':
          description: Conflict
        '500':
          description: Internal server error
components:
  schemas:
    UpdateUserResponse:
      type: object
      properties:
        header:
          $ref: '#/components/schemas/ResponseHeader'
      required:
        - header
    GetUserResponse:
        type: object
        properties:
          header:
            $ref: '#/components/schemas/ResponseHeader'
          user:
            $ref: '#/components/schemas/User'
        required:
          - header
          - user
    UserLoginResponse:
        type: object
        properties:
          header:
            $ref: '#/components/schemas/ResponseHeader'
          user:
            $ref: '#/components/schemas/User'
        required:
          - header
          - user
    RegisterUserResponse:
      type: object
      properties:
        header:
          $ref: '#/components/schemas/ResponseHeader'
        user:
          $ref: '#/components/schemas/User'
      required:
        - header
        - user
    ResponseHeader:
      type: object
      properties:
        success:
          type: boolean
          description: Boolean to denote whether response is OK or not.
        messages:
          description: Array of error message(s).
          type: array
          items:
            type: string
      required:
        - success
        - messages
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        phone_number:
          type: string
          description: User's phone number.
        full_name:
          type: string
          description: User's full name.
        password:
          type: string
          description: User's password.